General Code Fixes and Clean Up
Added CanAttack() method to Attacks script for improved code readability.
Made attack functions private.
Improved building damage system:
- Added private float field representing player’s scaled attack damage to Attacks
-Renamed ScaleAttackRate() to Scale(), and made it scale player’s attack damage as well
- Turned building’s health and scale fields into floats
- Made PlayerController’s Grow() function take the building’s scale in the form of a float rather than an int
Removed Attacks field from ComboList script.
Made player movement framerate independent:
- Added private inputX and inputY float fields to PlayerController
        - Gathered these inputs from Input system in Update()
- Set player rigidbody velocity in FixedUpdate()
Replaced PlayerController’s ScaleMovementSpeed() with Scale(), which scales the player’s speed and size their current scale.
Reworded many comments.
Renamed UIButtonBehavior interface to fit naming conventions.

+++ Assets/Scripts/BuildingBehavior.cs
@@ -4,13 +4,13 @@ using UnityEngine;
 // ReSharper disable All
 
 /// <summary>
-/// Behavior script for Building objects. Buildings
-/// can take damage from players and be destroyed.
+/// Script for handling the behavior of Building objects.
+/// Buildings can take damage from players and be destroyed.
 /// </summary>
 public class BuildingBehavior : MonoBehaviour
 {
-    public int health;
-    public int scale;
+    public float health;
+    public float scale;
 
     private SpriteRenderer sr;
 
@@ -25,8 +25,8 @@ public class BuildingBehavior : MonoBehaviour
     /// player if the player's scale is greater
     /// than or equal to that of the building.
     /// 
-    /// Calls DestroyBuilding() if health 
-    /// reaches zero.
+    /// Destroys the building if health reaches
+    /// 0.
     /// </summary>
     /// <param name="damage">
     /// Damage to be taken by the building.
@@ -35,7 +35,7 @@ public class BuildingBehavior : MonoBehaviour
     /// PlayerController of the player inflicting
     /// damage.
     /// </param>
-    public void TakeDamage(int damage, PlayerController pc)
+    public void TakeDamage(float damage, PlayerController pc)
     {
         pc.GetShrinkTimer().Reset(); //reset shrink timer
